Move cat_annotator fixture to conftest
